IGNITE-26012 Split BinaryTupleComparatorTest
